% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOI_plot.R
\name{calculate_marker_moi}
\alias{calculate_marker_moi}
\title{Calculate Marker-Specific MOI}
\usage{
calculate_marker_moi(genotypedata, marker_pattern = "_\\\\d+$")
}
\arguments{
\item{genotypedata}{A data frame with \code{Sample.ID}, \code{Site}, and marker columns.}

\item{marker_pattern}{A regex pattern to identify marker columns.}
}
\value{
A data frame summarizing MOI for each sample and marker.
}
\description{
Calculate Marker-Specific MOI
}
