% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_plot.R
\name{generate_allele_frequency_plot}
\alias{generate_allele_frequency_plot}
\title{Generate and save an allele frequency distribution plot}
\usage{
generate_allele_frequency_plot(
  raw_data_df,
  site_name = NULL,
  output_folder,
  plots_per_row = 4,
  plot_width = 12,
  plot_height = 10
)
}
\arguments{
\item{raw_data_df}{A data frame containing raw genotyping data. For a site-specific
plot, this should be filtered for one site. For an overall plot, it should
contain data from all sites.}

\item{site_name}{The name of the site for the plot title and filename.
If \code{NULL} (the default), an "Overall" plot is generated.}

\item{output_folder}{The directory where the plot PDF will be saved.}

\item{plots_per_row}{The number of marker plots to display per row.}

\item{plot_width}{The width of the output PDF file.}

\item{plot_height}{The height of the output PDF file.}
}
\description{
This function creates a faceted bar plot showing the frequency of each allele
length for Day 0 vs. Day of Failure samples. It can generate plots for a
single site or an overall plot for all sites combined.
}
