% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOI_plot.R
\name{generate_marker_moi_plot}
\alias{generate_marker_moi_plot}
\title{Generate and Save an MOI Plot}
\usage{
generate_marker_moi_plot(
  moi_data,
  output_folder,
  filename = "moi_per_marker_by_site.pdf"
)
}
\arguments{
\item{moi_data}{A data frame of MOI values, typically from \code{calculate_marker_moi}.}

\item{output_folder}{Path to the directory for saving the plot.}

\item{filename}{The name for the output PDF file.}
}
\value{
The ggplot object of the MOI plot.
}
\description{
Generate and Save an MOI Plot
}
