% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalReBay-package.R
\docType{package}
\name{MalReBay-package}
\alias{MalReBay-package}
\title{MalReBay: Bayesian classification for malaria recurrences}
\description{
\if{html}{\figure{logo.svg}{options: style='float: right' alt='logo' width='120'}}

Classification of malaria infection recurrence using a Bayesian algorithm implemented in Stan for efficient sampling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://swisstph.github.io/MalReBay}
  \item \url{https://github.com/swisstph/MalReBay}
  \item Report bugs at \url{https://github.com/swisstph/MalReBay/issues}
}

}
\author{
\strong{Maintainer}: Veronica Adhiambo Ochieng \email{veronica.adhiambo@aims.ac.rw}

Authors:
\itemize{
  \item Veronica Adhiambo Ochieng \email{veronica.adhiambo@aims.ac.rw}
  \item Mateusz Plucinski
  \item Monica Golumbeanu \email{monica.golumbeanu@swisstph.ch}
}

}
\keyword{internal}
