% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import Genotyping Data from Excel}
\usage{
import_data(filepath, marker_filepath = NULL, verbose = TRUE)
}
\arguments{
\item{filepath}{The full path to the input Excel file.}

\item{marker_filepath}{Path to Excel file containing marker metadata (optional if marker_info sheet is present)}

\item{verbose}{Logical. If TRUE, prints progress and data-cleaning messages.}
}
\value{
A list containing the imported data.
}
\description{
Reads data from a specified Excel file, automatically detecting
the data type and separating sheets into a structured list.
}
