% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MalReBay-package.R
\docType{package}
\name{MalReBay-package}
\alias{MalReBay}
\alias{MalReBay-package}
\title{MalReBay: Bayesian classification for malaria recurrences}
\description{
Classification of malaria infection recurrence using a Bayesian algorithm.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://SwissTPH.github.io/MalReBay}
  \item \url{https://github.com/SwissTPH/MalReBay}
  \item Report bugs at \url{https://github.com/SwissTPH/MalReBay/issues}
}

}
\author{
\strong{Maintainer}: Veronica Adhiambo Ochieng \email{veronica.adhiambo@aims.ac.rw}

Authors:
\itemize{
  \item Mateusz Plucinski
  \item Monica Golumbeanu \email{monica.golumbeanu@swisstph.ch}
}

}
\keyword{internal}
