% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot_markers_diversity}
\alias{plot_markers_diversity}
\title{Generate and Save Diversity Pie Charts}
\usage{
plot_markers_diversity(
  genotypedata,
  data_type,
  marker_info = NULL,
  output_folder,
  filename_prefix = "diversity"
)
}
\arguments{
\item{genotypedata}{A dataframe containing the genotyping data.}

\item{data_type}{A string, either "length_polymorphic" or "ampseq".}

\item{marker_info}{A dataframe with information about the markers, required
if \code{data_type} is "length_polymorphic".}

\item{output_folder}{The path to the directory where the output PDF will be saved.}

\item{filename_prefix}{A string prefix for the output PDF filename.}
}
\value{
Invisibly returns the final combined ggplot object.
}
\description{
This function creates pie charts to visualize allele or haplotype
diversity at Day 0 and at the day of recurrence. It saves the combined plot
as an image.
}
